Update Swipe item view
