+++ Todoey/Controller/SwipeTableViewController.swift
@@ -1,0 +1,62 @@
+//
+//  SwipeTableViewController.swift
+//  Todoey
+//
+//  Created by Orlando Moraes Martins on 08/08/23.
+//
+
+import UIKit
+import SwipeCellKit
+
+
+class SwipeTableViewController: UITableViewController, SwipeTableViewCellDelegate {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        // Uncomment the following line to preserve selection between presentations
+        // self.clearsSelectionOnViewWillAppear = false
+        
+        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
+        // self.navigationItem.rightBarButtonItem = self.editButtonItem
+    }
+    
+    //MARK: - table view DataSource methods
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        
+        let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath) as! SwipeTableViewCell
+        
+        cell.delegate = self
+        
+        return cell
+    }
+    
+        
+    func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath, for orientation: SwipeActionsOrientation) -> [SwipeAction]? {
+        guard orientation == .right else { return nil }
+
+        let deleteAction = SwipeAction(style: .destructive, title: "Delete") { action, indexPath in
+            
+            self.updateModel(at: indexPath)
+
+            print("delete son")
+        }
+
+        // customize the action appearance
+        deleteAction.image = UIImage(named: "Trash-icon")
+
+        return [deleteAction]
+    }
+    
+    func tableView(_ tableView: UITableView, editActionsOptionsForRowAt indexPath: IndexPath, for orientation: SwipeActionsOrientation) -> SwipeOptions {
+        var options = SwipeOptions()
+        options.expansionStyle = .destructive
+        options.transitionStyle = .reveal
+        return options
+    }
+    
+    func updateModel(at indexPath: IndexPath) {
+        // Update our date model
+    }
+    
+}
